Tweak _on_recommended_state_changed log message
I was looking at the logs to verify this behavior and I thought this
message could be a little clearer. It didn't mention the tab being acted
on (important when tying together logs across signal firings and with
multiple tabs open) and I switched it to be in logfmt instead of a
whole sentence for ease of parsing.

I also added a `if current == new: return` check, again because of the
logs. Having log messages that won't have any effect probably aren't
needed, and I was confusing them with the important logs.

Unfortunately exciting early means we don't connect new signals, so next
time we come through there is nothing to disconnect. Which makes the
`disconnect()` call grumpy. TypeError is not a super meaningful error to
be throwing here, but I assume that's a result of the python and c++
integration. If the signal was connected to a slot we could
unambiguously disconnect just that one slot, but it's anonymous
currently so we'll see how this goes.