Support disabling a lifecycle state by setting a delay of -1
This is inspired by previous iterations of this feature. Discarding a
page can lose information on the page, for example, if you go to
duckduckgo.com and put some text in the search box, then switch to a
different tab and wait for the ddg tab to be discarded, when you switch
back to the ddg tab the search box will be empty.

Until a point where we have more sophisticated controls of what sites
these transitions apply to, it may be useful for users to disable the
discard state completely. I know that's how I've had the lazier loading
implementation set up for years.

Negative numbers would previously have raised a warning on every state
transition where the time complained that negative numbers weren't
supported.