Don't apply lifecycle rules to pinned tabs
I want to make sure tabs in the background that might show notifications
or update their favicons or whatever can keep running if I want them to,
and not get frozen. Previous iterations of this feature have re-used the
pinned state to demark tabs that shouldn't be frozen. So that's what
I've implemented here. What could go wrong?

If kept the same "do act on the state change if we are transitioning to
an active state" check from the top of the function for consistency, but
neither check is needed. Qt will put tabs into the Active state when
they are shown without anything needed from our code.

I also haven't implemented a lifecycle change when you un-pin a tab.
It'll get picked up when you switch away from the tab.