Update min/max values for lifecycle settings
It doesn't make sense to allow setting these to negative values, so lets
help the user out by refusing to do so if they try.

But leave `-1` as an allowed value for the "discard" option so users can
have "freeze" enabled but turn the discard state off.

Remove the mention of -1 from the freeze setting since we don't allow
that anymore.